requirements.txt
environs==9.5.0
python-telegram-bot==11.1.0
notifiers==1.3.3
vk-api==11.9.9